+++ ClickerHeroesTrackerWebsite/ClickerHeroesTrackerWebsite/Models/SaveData/ItemData.cs
@@ -28,7 +28,7 @@ namespace ClickerHeroesTrackerWebsite.Models.SaveData
         /// Gets or sets the first item bonus level.
         /// </summary>
         [JsonProperty(PropertyName = "bonus1Level")]
-        public int? Bonus1Level { get; set; }
+        public double? Bonus1Level { get; set; }
 
         /// <summary>
         /// Gets or sets the second item bonus type.
@@ -40,7 +40,7 @@ namespace ClickerHeroesTrackerWebsite.Models.SaveData
         /// Gets or sets the second item bonus level.
         /// </summary>
         [JsonProperty(PropertyName = "bonus2Level")]
-        public int? Bonus2Level { get; set; }
+        public double? Bonus2Level { get; set; }
 
         /// <summary>
         /// Gets or sets the third item bonus type.
@@ -52,7 +52,7 @@ namespace ClickerHeroesTrackerWebsite.Models.SaveData
         /// Gets or sets the third item bonus level.
         /// </summary>
         [JsonProperty(PropertyName = "bonus3Level")]
-        public int? Bonus3Level { get; set; }
+        public double? Bonus3Level { get; set; }
 
         /// <summary>
         /// Gets or sets the fourth item bonus type.
@@ -64,6 +64,6 @@ namespace ClickerHeroesTrackerWebsite.Models.SaveData
         /// Gets or sets the fourth item bonus level.
         /// </summary>
         [JsonProperty(PropertyName = "bonus4Level")]
-        public int? Bonus4Level { get; set; }
+        public double? Bonus4Level { get; set; }
     }
 }
